+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2-preview] - 2020-03-11
+### Added
+- LocalTimeFactor for independent time manipulation
+
+### Removed
+- IsBackToTheFuture component is removed
+
+### Fixed
+- LocalTimeSlider now working
+
 ## [0.0.1] - 2020-03-11
 ### Added
 - LocalTime for ECS
@@ -12,4 +22,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Some useful Authoring scripts
 - DestoryMeshRendererOnStart for Hybrid ECS
 - LocalToWorld2TranslationAndRotation for CopyTransformFromGameObject on Hybrid ECS
-